requirements.txt
absl-py==1.3.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
azure-core==1.26.2
azure-identity==1.12.0
azure-storage-blob==12.14.1
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
contourpy==1.0.6
cryptography==39.0.0
cycler==0.11.0
decorator==4.4.2
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.1.4
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
imutils==0.5.4
isodate==0.6.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
moviepy==1.0.3
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==22.0
pafy==0.5.5
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pooch==1.6.0
portalocker==2.7.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
replicate==0.4.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
soundfile==0.11.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
twilio==7.16.1
typing_extensions==4.4.0
urllib3==1.26.13
webcolors==1.12
Werkzeug==2.2.2
wrapt==1.14.1
youtube-dl==2021.12.17
zipp==3.11.0
